video_analysis/requirements.txt
cached-property==1.5.2
func-timeout==4.3.5
grpcio==1.36.1
grpcio-tools==1.36.1
scipy==1.6.1
tabulate==0.8.7
easyocr